rename: 공모전 및 대외활동 컴포넌트 containers로 이동
Related to: #32
