+++ src/containers/activity/CategoryPanel.tsx
