+++ src/containers/contest/CategoryPanel.tsx
