+++ src/containers/contest/ResetButton.tsx
